requirements.txt
Flask==2.3.2
gunicorn==21.2.0
torch==2.1.0
torchaudio==2.1.0
demucs==4.0.1
noisereduce==3.0.0
scipy
numpy
# --- New Libraries ---
firebase-admin==6.2.0
google-cloud-storage==2.10.0
celery==5.3.1
redis==4.6.0
Flask-Babel==3.0.0
google-cloud-secret-manager
